chore: moving lib to utils inside the test folder and git ignore vscode config changes
