+++ .gitignore
@@ -22,6 +22,7 @@ lerna-debug.log*
 .DS_Store
 .fleet
 .idea
+.vscode
 
 # Local env files
 .env
