+++ tests/e2e/homepage.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from '../../lib/fixtures/base.fixture'
+import { expect, test } from './utils/fixtures/base.fixture'
 
 test.describe('Home Page/Landing Page', () => {
   // Check the Landing Page Main Elements
