+++ tests/e2e/utils/fixtures/statistics.fixture.ts
