+++ tests/e2e/utils/pages/privacy.page.ts
