+++ tests/e2e/utils/pages/statistics.page.ts
