+++ tests/e2e/utils/pages/terms.page.ts
